+++ spacy/cli/converters/iob2json.py
@@ -2,8 +2,8 @@
 from __future__ import unicode_literals
 from cytoolz import partition_all, concat
 
+from .._messages import Messages
 from ...compat import json_dumps, path2str
-from ...errors import Messages
 from ...util import prints
 from ...gold import iob_to_biluo
